feat(layout): add footer
Add a component for the footer.
Align it so that it always appears at the bottom of the page, even if
the content is shorter than full page height.

This commit also adds a `Layout` component that adds the header and the
footer automatically and accepts the main content of the page.