refactor(indexes): use content file paths relative to repo root
On Vercel, using `import.meta.env` or `__dirname` is only valid for the
duration of the build. After the files are built and run on Edge
Functions, their `import.meta.env` and `__dirname` remain the same, but
the files are in a totally different directory. `import.meta.env` or
`__dirname` no longer exist at that point.

Using `process.cwd` is the only correct way to get the path to the
repository root. It works in development, in production when using
`npm run build`, and on Vercel.