fix(article): include content files and indexes in Vercel
Vercel deployments aggressively exclude files which are not imported or
otherwise determined as necessary by Next (using their tracing
utilites) [0]. This happens to exclude the content files and indexes,
which are neeced during Incremental Static Generation on Edge Functions.

`unstable_includeFiles` is a way to tell Vercel to keep those files
in the Edge Functions.

[0]: https://nextjs.org/docs/advanced-features/output-file-tracing#caveats