refactor(styles): use px for distances, rems for font-sizes
Change the approach used to specify values in CSS.

* For lengths, like widths, heights, gaps, radiuses, paddings, margins,
  use `px`.
* For font sizes, use `rem`.

This way changing the root font-size, either in page CSS or in the
browser settings, affects only the sizes of text on the page. Distances,
margins, paddings, are unaffected. They are affected by zooming in/out
of the page.

This makes it possible to only increase/decrease the font-size without
zooming in/out. Previously, changing the font-size resulted in the same
effect as if the page was zoomed in.

Moreover, the `rem` utility was renamed to `pxToRem`.
`pxToRem` and `spacing` are now accessible on the Emotion `theme` and
should be preferred over regular imports.