fix: next/link usage
Next 13 changed the way next/link should be used. The new behavior is
for next/link to render the `a` tag. Thus, it should not receive the `a`
tag as children. If that happens, the `legacyBehvaior` prop should be
provided.

This commit adjusts links to no longer render the duplicate `a` tag or
enable `legacyBehvaior` if needed.