fix: next/image usage
The new next/image component is no longer responsive by default. For
images with width and height, the image will have that dimensions on the
page. `max-width` and other responsive styles need to be applied by
application code.

This commit adjusts existing uses of `next/image` to make the images
responsive. The about page hero image was also reworked to rely more on
aspect-ratio and object-fit styles.